+++ include/Utilities.h
@@ -9,9 +9,9 @@ the aspect ratio
  * @param bbox
  * @return z
  */
-    Eigen::VectorXd convert_bbox_to_z(Eigen::VectorXd bbox);
-    Eigen::VectorXd speed_direction(Eigen::VectorXd bbox1, Eigen::VectorXd bbox2);
-    Eigen::VectorXd convert_x_to_bbox(Eigen::VectorXd x) ;
-    Eigen::VectorXd k_previous_obs(std::unordered_map<int, Eigen::VectorXd> observations_, int cur_age, int k);
+    Eigen::VectorXf convert_bbox_to_z(Eigen::VectorXf bbox);
+    Eigen::VectorXf speed_direction(Eigen::VectorXf bbox1, Eigen::VectorXf bbox2);
+    Eigen::VectorXf convert_x_to_bbox(Eigen::VectorXf x) ;
+    Eigen::VectorXf k_previous_obs(std::unordered_map<int, Eigen::VectorXf> observations_, int cur_age, int k);
 }// namespace ocsort
 #endif//OC_SORT_CPP_UTILITIES_H
